pyproject.toml
[project]
name = "github-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "dotenv>=0.9.9",
 "fastmcp>=2.11.0",
 "genai>=2.1.0",
 "google>=3.0.0",
 "google-genai>=1.28.0",
 "google-generativeai>=0.8.5",
 "pygithub>=2.7.0",
 "python-dotenv>=1.1.1",
]


